build: Add model hack to bypass SafetyNet hardware attestation
This adds a new variable, PRODUCT_SAFETYNET_MODEL_HACK, that products
can set to "true" to bypass SafetyNet hardware attestation and make it
revert back to basic software checks. It works by appending an invisible
character to the model name so that SafetyNet doesn't recognize the
device as eligible for hardware attestation, but users can't notice the
difference because it's invisible.

Change-Id: I3bb408a66a166d67da660e773a4202b3651dd84f