build: Only use RRO for framework on mainline products
We need this to be able to overlay packages with prebuilt vendors that
overlay SystemUIGoogle and SettingsGoogle.

Ideally, we would disable RRO for all packages for performance
(especially framework-res), but there's a bug in our framework overlay
that causes quick settings to have the wrong background color in dark
mode.

Change-Id: I24a5e7fade70d89a20516f0a7f25cf40254d1403